Capture `stdout` and `stderr` even if `cmd.Run` fails
In its current form, if `cmd.Run` fails, then the function immediately
exits with no reason given for the termination. This makes debugging
scripts hard to understand as no indication is given for the failure,
other than a failure occured.

By trapping the error returned from `cmd.Run`, we first process the
buffers for both `stdout` and `stderr` and set these to the `dxr`.

Additionally we capture a truncated `stderr` from the command error
which can help point to where the command failed, setting this as the
Fatal error message returned as an event.

Signed-off-by: Martin Proffitt <mproffitt@choclab.net>